edge-modules/spike-predictor/requirements.txt
flask
numpy
scikit-learn
joblib
azure-iot-device